requirements.txt
coloredlogs==15.0.1
colorzero==2.0
gpiozero==2.0
humanfriendly==10.0
paho-mqtt==1.6.1
RPi.GPIO==0.7.1
RPLCD==1.3.1
smbus==1.1.post2